+++ crates/nostr/src/nips/mod.rs
@@ -15,7 +15,6 @@ pub mod nip05;
 pub mod nip06;
 pub mod nip09;
 pub mod nip10;
-#[cfg(all(feature = "std", feature = "nip11"))]
 pub mod nip11;
 pub mod nip13;
 pub mod nip15;