+++ crates/nostr/src/prelude.rs
@@ -37,7 +37,6 @@ pub use crate::nips::nip05::{self, *};
 pub use crate::nips::nip06::{self, *};
 pub use crate::nips::nip09::{self, *};
 pub use crate::nips::nip10::{self, *};
-#[cfg(all(feature = "std", feature = "nip11"))]
 pub use crate::nips::nip11::{self, *};
 pub use crate::nips::nip13::{self, *};
 pub use crate::nips::nip15::{self, *};