+++ CHANGELOG.md
@@ -36,10 +36,12 @@
 - nostr: remove `TagStandard::Delegation` and `TagKind::Delegation` (https://github.com/rust-nostr/nostr/pull/929)
 - nostr: remove `nip05` cargo feature (https://github.com/rust-nostr/nostr/pull/936)
 - nostr: convert `nip05` module to be I/O-free (https://github.com/rust-nostr/nostr/pull/936)
+- nostr: convert `nip11` module to be I/O-free (https://github.com/rust-nostr/nostr/pull/950)
 - nostr: convert `nip96` module to be I/O-free (https://github.com/rust-nostr/nostr/pull/935)
 - connect: remove `NostrConnect::get_relays` (https://github.com/rust-nostr/nostr/pull/894)
 - database: merge traits into `NostrDatabase` (https://github.com/rust-nostr/nostr/pull/916)
 - database: remove `NostrDatabase::has_coordinate_been_deleted` (https://github.com/rust-nostr/nostr/pull/917)
+- pool: remove NIP-11 support (https://github.com/rust-nostr/nostr/pull/950)
 - mls: changed return type of `NostrMls::add_members` and `NostrMls::self_update` (https://github.com/rust-nostr/nostr/pull/934)
 - mls: changed return type of all group and message methods to return Events instead of serialized MLS objects. (https://github.com/rust-nostr/nostr/pull/940)